% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{cat_summarise}
\alias{cat_summarise}
\title{Summarise the values of a numerical variable for each group within a
categorical variable and return the results as a tibble}
\usage{
cat_summarise(data, cat, num, na.rm = FALSE, clean_names = TRUE)
}
\arguments{
\item{data}{A data frame containing the vector to summarise.}

\item{cat}{The name of a column in \code{data} which is a categorical vector
of discrete values for which frequencies will be calculated.}

\item{num}{The name of a column in \code{data} which is a numerical vector
that will be summarised for each group.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}

\item{clean_names}{A boolean indicating whether the column names of the
results tibble should be cleaned, so that any column names produced from
data are converted to snake_case. The default is TRUE.}
}
\description{
The results are sorted on the values of the categorical variable i.e.
the variable specified with \code{cat}. If this variable is a character
vector it will be sorted alphabetically, and if it is a factor it will be
sorted in the order of its levels.
}
