% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcount.R
\name{cat_vcount}
\alias{cat_vcount}
\title{Count the frequency of discrete values in a categorical vector}
\usage{
cat_vcount(
  cat,
  name = NULL,
  by = "number",
  order = ifelse(by == "category", "asc", "desc"),
  na.rm = FALSE
)
}
\arguments{
\item{cat}{A categorical vector of discrete values for which
frequencies will be calculated.}

\item{name}{The column name to use for the column of categories in the
results as a string. By default, the variable name of the input vector is
used.}

\item{by}{The column name by which to sort the results as a string. Must be
one of either "n" or "cat". The default is "count".}

\item{order}{The order in which to sort the results as a string. Must be
one of either "asc" or "desc". The default depends on the value of the
\code{by} argument. If \code{by} is "count" the default \code{order} is
"desc". If \code{by} is "category" the default \code{order} is "asc".}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}
}
\value{
A tibble showing the frequency of each value in the input vector.
}
\description{
Calculates the frequency of discrete values in a categorical vector and
returns the counts as a tibble. This function differs from \code{cat_count}
in that it operates directly on vectors. rather than on columns in
dataframes, which means it's less useful in pipelines but can handle a
wider range of inputs.
}
