% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{safe_mean}
\alias{safe_mean}
\title{Calculate \code{mean} but return NA rather than NaN when values are missing}
\usage{
safe_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numerical vector.}

\item{na.rm}{A boolean indicating whether to remove NAs.}
}
\value{
The mean of \code{x} or NA when values are missing.
}
\description{
This function is a drop-in replacement for \code{mean}, which is used in
\code{cat_summarise}. It returns NA rather than NaN when all values are NA.
}
\keyword{internal}
