% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbycat-package.R
\docType{package}
\name{tabbycat}
\alias{tabbycat}
\alias{tabbycat-package}
\title{tabbycat: Tabulate and summarise categorical data}
\description{
Functions for tabulating and summarising categorical variables.
}
\author{
\strong{Maintainer}: Oliver Hawkins \email{oli@olihawkins.com}

}
\keyword{internal}
