% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{cat_compare}
\alias{cat_compare}
\title{Calculate the frequency of discrete values in one categorical variable for
each group within another categorical variable and return the results as a
tibble}
\usage{
cat_compare(
  data,
  dist_cat,
  group_cat,
  na.rm = FALSE,
  clean_names = TRUE,
  only = ""
)
}
\arguments{
\item{data}{A dataframe containing the two variables of interest.}

\item{dist_cat}{The column name of a categorical variable whose distribution
will be calculated for each group in \code{group_cat}.}

\item{group_cat}{The column name of a categorical variable which will be
split into groups and the distrubtion of \code{dist_cat} calulated
for each group.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}

\item{clean_names}{A boolean indicating whether the column names of the
results tibble should be cleaned, so that any column names produced from
data are converted to snake_case. The default is TRUE.}

\item{only}{A string indicating that only one set of frequency columns
should be returned in the results. If \code{only} is either "n" or
"number", only the number columns are returned. If \code{only} is either
"p" or "percent", only the percent columns are returned. If \code{only} is
any other value, both sets of columns are shown. The default value is an
empty string, which means both sets of columns are shown.}
}
\value{
A tibble showing the distribution of \code{dist_cat} within each of
  the two exclusive groups in \code{group_cat}.
}
\description{
This function crosstabulates the frequencies of one categorical variable
within the groups of another.
}
\details{
The results are sorted on the values of the variable whose distribution is
shown in each column i.e. the variable specified with \code{dist_cat}. If
this variable is a character vector it will be sorted alphabetically, and if
it is a factor it will be sorted in the order of its levels.
}
