% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{cat_contrast}
\alias{cat_contrast}
\title{Calculate the frequency of discrete values in one categorical variable for
each of two mutually exclusive groups within another categorical variable
and returns the results as a tibble}
\usage{
cat_contrast(
  data,
  dist_cat,
  group_cat,
  group_name,
  na.rm = FALSE,
  clean_names = TRUE,
  only = ""
)
}
\arguments{
\item{data}{A dataframe containing the two variables of interest.}

\item{dist_cat}{The name of a categorical variable whose distribution should
be calculated for each group.}

\item{group_cat}{The name of a categorical variable that will be split
into two exclusive groups, one belonging to observations with a
particular value of that variable, and another containing all other
observations.}

\item{group_name}{The name of the group within \code{group_cat} that is
used to split the observations into two exclusive groups: those that are
in the group and those that are not in the group.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}

\item{clean_names}{A boolean indicating whether the column names of the
results tibble should be cleaned, so that any column names produced from
data are converted to snake_case. The default is TRUE.}

\item{only}{A string indicating that only one set of frequency columns
should be returned in the results. If \code{only} is either "n" or
"number", only the number columns are returned. If \code{only} is either
"p" or "percent", only the percent columns are returned. If \code{only} is
any other value, both sets of columns are shown. The default value is an
empty string, which means both sets of columns are shown.}
}
\value{
A tibble showing the distribution of \code{dist_cat} within each of
  the two exclusive groups in \code{group_cat}.
}
\description{
This function shows the distrbution of values within given a categorical
variable for one group within another categorical variable, and compares it
with the distribution among all observations not in that group. Its purpose
is to let you see quickly whether the distribution within that group differs
from the distribution for the rest of the observations.
}
