% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{cat_count}
\alias{cat_count}
\title{Count the frequency of discrete values in the column of a dataframe}
\usage{
cat_count(data, cat, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame containing the vector to summarise.}

\item{cat}{The name of a column in \code{data} which is a categorical vector
of discrete values for which frequencies will be calculated.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}
}
\value{
A tibble showing the frequency of each value in the input vector.
}
\description{
Calculates the frequency of discrete values in the column of a dataframe
containing and returns the counts as a tibble.
}
