% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{cat_count}
\alias{cat_count}
\title{Count the frequency of discrete values in the column of a dataframe and
return the results as a tibble}
\usage{
cat_count(data, cat, na.rm = FALSE, clean_names = TRUE, only = "")
}
\arguments{
\item{data}{A data frame containing a categorical vector to summarise.}

\item{cat}{The column name of a categorical variable for which frequencies
will be calculated.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}

\item{clean_names}{A boolean indicating whether the column names of the
results tibble should be cleaned, so that any column names produced from
data are converted to snake_case. The default is TRUE.}

\item{only}{A string indicating that only one of the frequency columns
should be returned in the results. If \code{only} is either "n" or
"number", only the number column is returned. If \code{only} is either
"p" or "percent", only the percent column is returned. If \code{only} is
any other value, both columns are shown. The default value is an empty
string, which means both columns are shown.}
}
\value{
A tibble showing the frequency of each value in the input vector.
}
\description{
This function differs from \code{cat_vcount} in that it operates on columns
in dataframes rather than directly on vectors, which means it is more useful
in pipelines but handles a narrower range of inputs.
}
\details{
The results are sorted in decending order of frequency.
}
