% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcount.R
\name{cat_vcount}
\alias{cat_vcount}
\title{Count the frequency of discrete values in a categorical vector}
\usage{
cat_vcount(cat, na.rm = FALSE, only = "")
}
\arguments{
\item{cat}{A categorical vector of discrete values for which
frequencies will be calculated.}

\item{na.rm}{A boolean indicating whether to exclude NAs from the results.
The default is FALSE.}

\item{only}{An optional argument indicating that only one of the frequency
columns should be returned in the results. If \code{only} is either "n" or
"number", only the number column is returned. If \code{only} is either
"p" or "percent", only the percent column is returned. If \code{only} is
any other value, both columns are shown. The default value is an empty
string, which means both columns are shown.}
}
\value{
A tibble showing the frequency of each value in the input vector.
}
\description{
Calculates the frequency of discrete values in a categorical vector and
returns the counts as a tibble. This function differs from \code{cat_count}
in that it operates directly on vectors. rather than on columns in
dataframes, which means it's less useful in pipelines but can handle a
wider range of inputs.
}
