% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabbycat.R
\docType{package}
\name{tabbycat}
\alias{tabbycat}
\title{tabbycat: Tabulate and summarise categorical data}
\description{
A small library of functions for exploring, tabulating, and summarising
categorical variables.
}
